完善 /upload 完成 /addLike

+++ picture-service/src/main/java/com/a205/mybed/pictureservice/pojo/Picture.java
@@ -1,6 +1,7 @@
 package com.a205.mybed.pictureservice.pojo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Picture implements Serializable {
@@ -12,7 +13,7 @@ public class Picture implements Serializable {
 
     private Date uploadTime;
 
-    private Integer size;
+    private BigDecimal size;
 
     private Date expireTime;
 
@@ -56,11 +57,11 @@ public class Picture implements Serializable {
         this.uploadTime = uploadTime;
     }
 
-    public Integer getSize() {
+    public BigDecimal getSize() {
         return size;
     }
 
-    public void setSize(Integer size) {
+    public void setSize(BigDecimal size) {
         this.size = size;
     }
